Add an error when saving a package if the package header gets too large which would cause one or more of the offsets stored in the FPackageFileSummary to truncate.
#rb Francis.Hurteau
#rnx
#preflight 628364050039ea57a52d6989

- The offsets and TotalHeaderSize value in FPackageFileSummary are all stored as int32 and although it is very unlikely that the package header would ever be large enough that we would need to store these values as 64 bit, we do support 64bit length files and if the header was large enough we could end up with truncated values in the summary which would cause the package to probably crash the editor when loaded.
- Since all of these values are assigned during ::WritePackageHeader we don't need to add a check every time one is assigned, we can have a single check at the end of writing out the header. Technically it might be faster to fail earlier when writing out the header, but it is unlikely that this condition will ever trigger and adding bespoke checks on every assignment would either bloat the code and make it harder to read or require a macro.
- I opted for a log error and not a user facing error because realistically if this occurs there is nothing an end user can do to fix it other than by making a bug report.

[CL 20257672 by paul chipchase in ue5-main branch]

+++ Engine/Source/Runtime/CoreUObject/Private/UObject/SavePackage/SavePackageUtilities.cpp
@@ -1601,7 +1601,7 @@ void SaveThumbnails(UPackage* InOuter, FLinkerSave* Linker, FStructuredArchive::
 				FObjectFullNameAndThumbnail& CurObjectThumb = ObjectsWithThumbnails[ CurObjectIndex ];
 
 				// Store the file offset to this thumbnail
-				CurObjectThumb.FileOffset = Linker->Tell();
+				CurObjectThumb.FileOffset = (int32)Linker->Tell();
 
 				// Serialize the thumbnail!
 				FObjectThumbnail* SerializableThumbnail = const_cast< FObjectThumbnail* >( CurObjectThumb.ObjectThumbnail );
@@ -1611,7 +1611,7 @@ void SaveThumbnails(UPackage* InOuter, FLinkerSave* Linker, FStructuredArchive::
 
 			// Store the thumbnail table of contents
 			{
-				Linker->Summary.ThumbnailTableOffset = Linker->Tell();
+				Linker->Summary.ThumbnailTableOffset = (int32)Linker->Tell();
 
 				// Save number of thumbnails
 				int32 ThumbnailCount = ObjectsWithThumbnails.Num();
@@ -2166,7 +2166,7 @@ void SaveWorldLevelInfo(UPackage* InOuter, FLinkerSave* Linker, FStructuredArchi
 	
 	if(FWorldTileInfo* WorldTileInfo = InOuter->GetWorldTileInfo())
 	{
-		Linker->Summary.WorldTileInfoDataOffset = Linker->Tell();
+		Linker->Summary.WorldTileInfoDataOffset = (int32)Linker->Tell();
 		Record << SA_VALUE(TEXT("WorldLevelInfo"), *WorldTileInfo);
 	}
 }
@@ -2230,7 +2230,7 @@ namespace AssetRegistry
 		check(!BinaryArchive.IsTextFormat());
 
 		// Store the asset registry offset in the file and enter a record for the asset registry data
-		Linker->Summary.AssetRegistryDataOffset = BinaryArchive.Tell();
+		Linker->Summary.AssetRegistryDataOffset = (int32)BinaryArchive.Tell();
 		FStructuredArchiveRecord AssetRegistryRecord = ParentRecord.EnterField(TEXT("AssetRegistry")).EnterRecord();
 
 		// Offset to Dependencies
